requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Data sources
dukascopy-python>=0.1.0
python-binance>=1.0.19

# Storage backends
duckdb>=0.8.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Configuration management
python-dotenv>=1.0.0

# UI for Colab
ipywidgets>=8.0.0

# Utilities
requests>=2.28.0
lzma 